requirements.txt
chromadb
groq
python-dotenv
pdfplumber
Pillow
langchain-text-splitters
sentence-transformers
streamlit
playwright
plotly